Add manual tracing support
For a long time now we had the plan to provide better introspection
capabilities in the form of tracing. Tracing captures additional context
that is vital for understanding emitted messages in an async program. We
have finally come around to investigating possible options on how to go
about that.
In a nutshell, the situation is bad. There are two main tracing crates,
tracing and async-log. Both of which are not very suitable in their
present form.
Async-log, seemingly coming from the asynd-std guys, doesn't bring much
value to the table to begin with. The good thing is that it uses the log
crate and so our existing infrastructure surrounding it will "just
work". However, we have seen a pull request adding async-std as a
dependency, making this an extremely dangerous choice. That, combined
with a questionable value-add, disqualify it.
The tracing crate is fairly sophisticated. So sophisticated, in fact,
that it took a day or so to understand how to work with it and convince
an application to log traces. There are various was to enable logging:
- the "log" feature enables rudimentary logging of all events/spans;
  this logging is so rudimentary that it isn't even useful: context is
  not included properly and so the value add over mere logging is close
  to zero
- the tracing-log crate comes with some support for env_logger-like
  initialization -- but that didn't really work as intended
- tracing-subscriber comes with a nicely configurable formatter that
  emits logs (with span contexts) to stdout; it can be configured
  env_logger-style

From the above three possibilities only tracing-subscriber is a viable
option. However, we decided against it because it is not straight
forward to enable tracing in tests. While test-env-log can be extended
to do so (and that works), really the initialization should happen only
once and not for every test. That is not a huge issue, but it is not
great either. The crate also comes with a shit ton of dependencies
(direct & indirect). Lastly, hyper uses log anyway and so by using a
tracing (the crate) based solution we effectively duplicate
functionality to a certain degree.
Because of all these issues and because we our requirements are very
minimal, we just roll our own tracing solution. In a nutshell, we create
custom span types (one for issuing request and one for stream
subscription) and just log those using the normal logging
infrastructure.