Fix tracing for issued requests
As experimentation has shown, the tracing span that we create in the
Client's issue method has a problem: for some reason it is not closed
properly. The result of this bug is that each subsequent request has the
previously issued request's context as the parent. This in turn leads to
hard to understand traces.
The problem is not fully understood at this point, but this change works
around it by using the instrument macro, which takes care of creating
the span for the annotated function/method. This approach does not have
the problem.