Move time_util module into root of crate
This change moves the time_util module out of the api module and inside
the root of the crate. In the future we will introduce another top-level
module that will make use of this functionality. There is no point in
claiming that anything in this module is specific to the api module.