Rename SystemTime parsing functions
With an upcoming change we will require more time parsing functions. In
particular, while we currently can parse a SystemTime from certain
standardized formats, we do not support creating such an object from a
UNIX time stamp directly.
In preparation for the addition of this new functionality, this change
renames the existing parsing functions, appending "from_str" to their
names.