Add support for endpoint-specific URLs
So far we have treated all endpoint definitions to be relative to a
single URL. To that end we did not include the URL to use into the
endpoint at all, but just let clients deal with that. Unfortunately such
a scheme will not be sufficient in the future, where we want to work
with an endpoint that uses a different base URL.
To accommodate for such cases, the Endpoint trait from the http-endpoint
crate has been adjusted to include a method potentially returning an
endpoint specific base URL.
This change makes our Client struct work correctly with such endpoint
definitions.