Add support for serializing a SystemTime into RFC3339 time stamp
In addition to being able to deserialize a SystemTime from a UNIX time
stamp, we will also be required to serialize such a time. The desired
format, however, is not a mere UNIX time stamp but an RFC3339
formatted one.
This change adds support for performing such a task through a newly
introduced function, system_time_to_rfc3339.