Use serde_urlencoded for creating endpoint query strings
We have a few endpoints that require a query string to be used to
provide certain input. The way we create those query strings is by using
the url crate's form_urlencoded serializer. The problem with this
approach is that we are now using custom query key definitions inside
the Endpoint's query method, when ideally we could just piggy back on
serde's field annotations.
This change rectifies the situation by using the serde_urlencoded crate
to perform this job. We already had the fields in the request types
annotated with the key names and so all we need to do is to use a
different (and much more concise) way to serialize the query string.
Unfortunately doing so introduces a new error path that we are not
really equipped to handle. For now we just unwrap, which should be fine
given that we have full control over the input strings (preventing
problems such as UTF-8 errors).