Add support for changing an existing order
The Alpaca API has limited support for changing an existing order.
Limited refers to the fact that the symbol to trade or the side of the
trade cannot be changed. Other properties such as the quantity or limit
price can be amended after initial submission, however.
This change adds support for changing an order in such a way by issuing
a PATCH request to the orders endpoint. Note that those orders actually
appear with a status that was unknown to use so far: replaced. As such,
the order Status enum is extended with this new variant.