Introduce TradeStatus::to_order_status helper method
The events::TradeStatus and order::Status enums are almost -- but not
quite -- identical: the order::Status has an additional variant
indicating that an order has been accepted for bidding.
However, what this means is that we can always convert a
events::TradeStatus into an order::Status (while the converse is not
true).
Performing such a conversion may help clients eliminate some redundancy
with respect to trade update events. Hence, this change introduces a
method to the TradeStatus enum that allows for that.