Add support for listing orders based on their status
So far we have only support listing open orders, which is the API
default unless specified differently. However, we have come to the point
where it would be great to also list closed orders.
To support such a case, this change extends the OrdersReq struct with a
status parameter that defines the status of orders to return.