Introduce Replaced variant to TradeStatus enum
As it turns out the trade status we receive from Alpaca can also contain
a (undocumented) "replaced" variant. At this point it is very
questionable whether there really are any differences to an order
status. Unfortunately, at least the documentation states that there is a
different identifier used to indicate partial fills, which is why we
introduced a dedicated type to begin with.
In any case, in order to fix this problem this change introduces a
Replaced variant.