Fix inner tracing span for Client::subscribe
Investigation of traces has shown that the tracing of the manually
created span inside the Client::subscribe method is showing up in places
where it should not. This is a known problem we have fought with for
long, but never gotten to the root of. One workaround is to use the
instrument procedural macro.
Repeated investigation has now brought to light the issue: we were using
tracing's info_span for creation of the span, assuming that it is
futures aware. That, however, is not the case. To fix this issue we now
switch to using the future aware version, which unfortunately is quite a
bit more verbose.