Use proper time/date parsing functions instead of working with durations
We have a few locations in the test code that work with raw UNIX time
stamps (i.e., seconds since 1970-01-01). That is very unergonomic and
hard to understand.
By now we have exposed functions that can parse a string into a
SystemTime to the time-util crate and so with this change we switch over
to using them.