Make various quantities a u64
This change converts the Order and Position structs' quantity members
from a Num value to an u64. The quantity should always be an integer
value and there is little point in having the overhead of a Num. We had
thought about this step in the past, but the ambiguity of the Alpaca
documentation along with the fact that the numbers are actually provided
as strings kept us using Num. By now, the inconvenience of having a Num
in this place got too great.