Add support for negating "sides"
The Side enum we use for capturing which side an order or position is on
has exactly two values with "opposing" meaning. Client code may
reasonably want to negate a side and expect the result to be the other
one.
This change enables such use cases by implementing the Not trait for
both types.