Parse non-trade activity's quantity
In the past we encountered a bug in serde which prevented us from
parsing an optional non-trade activity's quantity. While the upstream
bug is still open, and the example program still fails, the code here in
question works as expected now. Hence, with this change we parse the
optional quantity of non-trade activities as well.