Fix failure of retrieve_order_by_id test case when market are open
When the markets are open the Alpaca API will necessarily behave a bit
differently than when they are closed. In the retrieve_order_by_id test
case, that exposes a problem when comparing the fields of two orders --
one as submitted and another one as returned after cancellation --
because not only can time stamps differ, but so can the status (when the
market is open the order may quickly transition into "accepted" state).

This is intended behavior and so with this change we simply no longer
compare the status fields of the two orders in question.