Parse quantity of position as signed first
As it turns out, the quantity of sell short orders is represented as a
negative value in the position object. That clashes with our attempt of
parsing it as a u64 directly.
To fix this problem, we decided against the most obvious solution of
simply switching to using i64 altogether. We consider a negative amount
as unintuitive and a break with the rest of the API (we are not ordering
negative amounts, even for short orders). Instead, we transparently drop
the sign of negative values instead. This approach has the additional
advantage that there are no type mismatches between quantities in order
and position objects, causing fewer error cases to be handled.