Add two more status variants to TradeStatus
As it turns out there are more variants that are undocumented and that
should be represented in the TradeStatus enum. At this point, the two
known missing ones are: the rejection of an order replacement and the
rejection of an order cancellation.
This change adds variants for those two events to the TradeStatus enum.