Do not log entire request object but just the body
Whenever we issue a request we log the entire request object. The
information logged includes the URL, which is already captured in the
containing span. That causes a lot of duplicate information to be
printed.
Really, the main purpose of this particular trace point was to provide
insight into what the body of the request looks like. Hence, with this
change we adjust it took log exactly that and not more.