Decrease tracing verbosity by one level
Given that this crate is more of an infrastructure crate, the
information emitted in traces is arguably too verbose. If we consider
INFO to be the default level, really only higher level primitives built
on top of this crate should use this level.
Hence, with this change be decrease the verbosity of all trace points
and spans by one level.