Introduce OrderReqInit type for default constructing OrderReq objects
The OrderReq type we have contains a lot of fields, most of which
actually -- at least in theory -- have sensible defaults. Unfortunately,
all of those fields need to be explicitly initialized when constructing
such an object.
With this change we introduce a way to abbreviate specification of
fields that actually do have defaults. Specifically, we introduce a new
type, OrderReqInit, that implements Default and which shares default
initializable fields of the same name with OrderReq. Users of the crate
not only can use this type to omit specifying all fields in OrderReq,
doing so also improves the forwards compatibility of code, because
relying this new type will remain default constructible, while new
fields may be added to OrderReq that will break existing code.