Support unknown enum variants more gracefully
Client software can potentially run into massive problems if Alpaca adds
a new variant for one of their statuses or anything else that we
represent as an enum. We have seen this happen with a new order status
being introduced, which resulted in all order retrievals failing.
Ultimately such behavior is unacceptable.
With this change we handle such variants gracefully, by falling back to
a dedicated 'Unknown' variant upon encounter. While we generally want to
treat such unrecognized variants as a bug, we cannot make them fail the
parsing of the enum, as that can have major implications.
We adjusted only those enums that we consider critical and at the same
time prone to such new additions.
Note that we do have a slight problem for types that are also meant to
be serializable, as they will not preserve the original status. For now
we have to live with this potential inconsistency. Long term we may
remove serialization support where it is not strictly needed, but we
introduced it because there was some need.