Remove TradeStatus::to_order_status method
With recent additions to the TradeStatus enum it became even more
apparent that this type is there to stay and will never be merged with
the order status. Furthermore, now that there are variants inside of
TradeStatus that do not correspond to an order status, it has become
even more questionable whether such a conversion functionality should be
provided to begin with, because that is considered a potential source of
errors.
With this change we remove the TradeStatus::to_order_status method from
the enum to not provide an easy way for performing such a lossy (and
presumed-to-be error-prone) conversion.