Simplify event decoding types
We require a set of types that map to objects contained in the reported
JSON in order to ultimately unwrap the innermost event object, which is
what we are interested in. As it turns out, we do have an unnecessary
unit type in there, which this change removes. Also, the inner (and
private) stream module is removed, as it is no longer needed.