Publicly export raw event stream
Some clients may not want to work with the stock event types we provide.
For example, in performance critical scenarios it may not be desired to
decode all fields of an event or only do so lazily. Or, clients may have
no use for the decoded data at all but just want to dump the raw JSON
data.
To enable these cases, this change exports the raw event stream as well,
allowing clients to do whatever they want with the received data. In
order to make this function usable, we also export the EventStream,
StreamType, and Event types.