Change stream and stream_raw to expect reference to ApiInfo object
While the stream and stream_raw are public functions, they are rather
clunky to use. In particular, the requirement for users to provide an
ApiInfo object is rather annoying, because while they are expected to
have a Client object, they can only retrieve a reference to the ApiInfo
object it was created with. Every new object instance requires a costly
clone.
This change reworks both functions to accept a reference to an ApiInfo
object. Note that internally we may still resort back to cloning. That's
probably fine, as the intention of this change is not to improve
performance but solely usability. Other changes could happen
transparently without affecting public APIs.
Also note that because of this desired change in the signature we have
to move the setting of the scheme deeper down into the streaming
functionality. And for that to work in conjunction with the tests we
need to make it conditional. That's not nice, and should probably be
cleaned up eventually.