Remove serialization support for account & trade events
This change removes all the definitions that allow for account & trade
events to be serialized again. This support for re-serialization has
never been considered perfect. For example, with the support of
catch-all unknown variants in certain enum types, if we were to
serialize them, the resulting object would be semantically different
than what we had initially. While that may be acceptable in certain
scenarios, that will blow up in others.
However, at the time introducing serialization support was an easy
solution to allowing a client application to emit the raw JSON data.
With recent changes we believe to have given users all the means to
build a different solution to solving this problem and as such we can
now safely remove this support.