Add support for paging of account activities
One long standing issue with the account activity retrieval has been
that it was limited to the most recent 100 transactions. That's not
necessarily a problem with low trading volumes, but with very high
volumes of trades or when live calculations are to be performed on the
activities this clearly is a limiting factor.
To fix this issue this change adds support for paging of account
activity data. Specifically, users can now provide a page size and a
pagination token (the ID of the last activity received) in the request.
With that they can iteratively page through all activities.