Remove Option wrapper from ActivityReq::types
The ActivityReq::types member has been introduced as an Option to easily
encode the fact that the desire is to not filter out certain activity
types. However, that opens up a possible state, Some with an empty Vec
inside, that logically does not make much sense.
To fix this issue, this change removes the Option wrapper from the
member. Given that the ActivityReq type is default initializable we do
not see a good reason for keeping the Option wrapper.