Remove Client::builder method
By now we effectively do no longer need the builder that we introduced
for the Client type: there is nothing to configure, because we only have
one property to set and we do so based on whether we compile for a test
or not.
To that end, this change removes the Client::builder method that allowed
for accessing the Builder type. We keep the builder type around for now,
but only for internal usage.