Use thiserror crate for defining error types
By now the various error types we use throughout the program have become
fairly complex and their specification/definition is a tedious and
repetitive endeavor.
To tame the beast somewhat, this change switches over to using the
thiserror crate for defining the error types. It allows for a fairly
declarative syntax that removes a lot of the repetition and helps us to
keep related properties in the same place. At the same time, it should
strictly be a compile-time expansion thing, not adding additional
symbols into a crated binary.
It does mean, however, that errors are reported/formatted a bit
differently. Specifically, we add more "meta" descriptions and rely on
the underlying error as the source to describe the actual problem.