Introduce ConversionError type
This change introduces a new error type that captures all types of
"conversion" errors related to HTTP endpoints: ConversionError. Most of
the conversion are from and to JSON and so this is one of the variants
it has. But the another one is a potential failure to urlencode a
string, as we do for certain query arguments. By having this variant we
fix a long standing issue where we have the (hopefully only
hypothetical) potential for a panic originating directly in the crate,
because we unwrap the result of URL query creation.