Wait for order to get canceled in list_orders test
We see the list_orders test failing occasionally in the CI pipeline,
failing the check that the supposedly-canceled order is no longer
listed.
The reason for this behavior seems to be that cancellation is happening
asynchronously, and until the corresponding event has been received, it
may not have happened.
Thus, to fix this test, with this change we make sure to wait for the
cancellation event before performing our checks.