Enable num-v04 feature for num-decimal
The num-* set of crates, which we indirectly rely on through our
dependency on the num-decimal crate, have had a few minor version bumps.
Because we indirectly embed types from it in our public interface we
have to be careful with what version of these crates we use such that
client code does not run into issues with type mismatches due to
different versions being used by different crates. At the same time, if
client code wants to use this crate but is already relying on a newer
version of the num-* crates and wants to interchange objects, such
conflicts can happen as well.
With this change we enable the num-v04 feature of the num-decimal crate,
which bumps the num-* crates consumed from 0.2 to 0.4. Doing so gets us
on par with the most up-to-date released version, which is the safest
path forward.