Adjust Client::issue to accept request input via reference
The Client::issue method takes ownership of the input parameters to use
for the request to issue. That, however, is actually unnecessary: none
of the internals actually require ownership of this data.
As such, this change adjusts the signature of the method to accept the
request input via a reference, eliminating the need to clone or
otherwise duplicate requests in certain client code.