Migrate api::v2::account to use DateTime<Utc>
This change migrates the api::v2::account module over to using
Chrono's DateTime<Utc> instead of a SystemTime. It step marks the first
step in performing this migration for the entire crate. There are almost
no actual benefits to working with a SystemTime and doing so arguably
impacts ergonomics negatively, because so few operations can be
performed on a SystemTime.