Migrate data::v1::bars to use DateTime<Utc>
This change continues the effort of migrating the crate from using
SystemTime to DateTime<Utc> and does so for the data::v1::bars module.
Note that we kept the Bar::time attribute as a SystemTime for now, as it
is delivered directly as a second value from upstream. Ultimately we
expect this module to be removed soon and so this decision should not
necessarily be regarded as final.