Use rustfmt to format the code base
The crate's code is formatted somewhat inconsistently in some places.
The reason being that over time both rustfmt evolved and our (not
checked-in) configuration for it changed slightly. The original plan was
to migrate lines as they are touched over time so as to circumvent the
need for a sweeping change. However, some lines are just rarely if ever
touched, so this process can take a while. Meanwhile, as more developers
are contributing, having an inconsistent style is becoming more and more
of an issue.
To that end, this change formats the entire code base with rustfmt. Not
too many locations need adjustment at this point, so the change luckily
is reasonable in size.