Drop dependency on time-util
By now we have converted all but one usage of the time-util crate: the
usage of the system_time_from_secs function for deserializing the
one SystemTime that we kept around in the Bar type.
With this change we reimplement this function directly in the crate,
allowing us to drop the dependency to time-util altogether.