Cache build artifacts in CI
CI end-to-end test times spiraled into almost surreal seeming realms. In
order to get them back under control, with this change we cache the
build artifacts from the target/ directory as used by cargo.
The cache is available not only between jobs of a single pipeline run,
but also across different pipeline runs. In our case, the cache will be
populated only by the build step. All subsequent jobs will merely read
from it. Because we have the cache available between jobs as well, we
can actually drop the usage of artifacts for transfer between stages.

Note that ideally we would cache cargo's home directory contents as
well, where the registry is stored, for example, but it appears that
doing so breaks rustup in unexpected ways. This problem may be specific
to the docker image we currently use, as we have no seen it elsewhere,
so chances are that we could retry after bumping to a newer version of
Rust.