Add support for notional orders
This change adds support for notional orders, i.e., those that specify a
mere dollar amount to purchase. Currently, Alpaca only supports this
order type for market orders that are valid for the current trading day.
In most cases a notional order will result in a fractional quantity to
be traded. As such, the account will also need to have fractional
trading enabled. Furthermore, given that we currently assumed integer
quantities of shares in any given position, this member was converted to
a Num.