Automatically remove stale files from CI cache
The change to cache the contents of our target/ directory in CI actually
had one unintended side effect: because we never clean the directory
potentially stale data can accumulate over time. In fact, we have
noticed dozens of test binaries being present in the directory after
just a few CI runs.
That is not really our intention and so with this change we find files
that have not been accessed as part of the build and just remove them.