Declare more types as non-exhaustive
We arguably still have a fairly unstable API overall, with a lot of
backwards incompatible changes. Part of it is the sheer size of the API
surface. But another is the fact that we even break compatibility when
just adding a member to most types (something that arguably happens
rather frequently due to upstream changes).
To mitigate this churn somewhat, with this change we declare more types
as non-exhaustive. We do that as per the following rules:
1) all *ReqInit types are by definition non-exhaustive
2) all struct response types are non-exhaustive; these are typically
   rather large objects for there is arguably little reason to destruct
   in an exhaustive manner