Add support for historical data v2 API
This change adds support for retrieving historical stock data using
Alpaca's v2 data API. v1 of the data API has been deprecated and should
no longer be used and so with this change we finally have a replacement.
Note that streaming of realtime data is a different matter and not yet
supported by the crate.