Add support for creating, retrieving, and deleting watchlists
This change adds support for basic management operations -- creating,
retrieving, and deleting -- for watchlists. A watchlist in Alpaca's sense
is basically a set of symbols associated with a name (and an ID) by
which it can be retrieved back. Note that we currently do not support
updating of watchlists, i.e., adding or removing symbols post creation.