Unhide documentation of NonTradeActivityImpl
When we originally added the NonTradeActivityImpl type we annotated it
with #[doc(hidden)], because it was an implementation detail. That's
still the case, but the attribute does not quite work as we intended it
to: the naive expectation was that we could still peek into the type
indirectly through the corresponding typedef, but as it turns out it is
hidden completely.
That's not what our intentions were and so with this change we remove
said attribute to make documentation available for the type.