Rename data::v2::stocks to data::v2::bars
This change renames the data::v2::stocks module to data::v2::bars. The
crate currently only supports stocks and as such the former name was not
very telling. Furthermore, in the future we want to add support for more
endpoints of the data API (also working for stocks) that cannot be
crammed into the original stocks module.