Introduce Bar struct and DataMessage enumeration
The Alpaca data API allows for streaming market data updates in real
time. Currently supported is streaming of quotes, trades, and bars
(i.e., aggregates).
In a first step towards enabling clients of the crate to use this API,
this change introduces the Bar struct that represents one such
aggregate. Such a bar is sent to us via a message that we represent with
the newly introduced enum DataMessage, which may support trades and
quotes in the future.