Introduce ParsedMessage abstraction
Similar to what we have for trade updates, this change introduces a
ParsedMessage typedef that represents the type that we have available
after parsing the "raw" websocket message. Furthermore, we implement the
Message trait for this type to classify such messages user-facing or
control messages, which will later allow us to later build a
subscription mechanism on top.