Introduce MarketData struct
The Alpaca real time data API allows users to subscribe to receive
different events: trades, quotes, and bars. This change introduces a
representation for a "request" to subscribe to such updates, called
MarketData. Note that currently we only support subscribing to aggregate
bars, but in the future we may opt to further support quotes and trades.