Introduce Request enum
In order to work with the Alpaca real time data API we need to send
requests for various operations. For example, after establishing a
connection we are required to authenticate with the API in order to
later be able to subscribe to updates and receive them.
This change introduces the necessary data definition representing such a
request, the Request enum.