Implement subscribe and unsubscribe for Subscription objects
This change implements the subscribe and unsubscribe methods of our
Subscription type. As their names imply, these can be used to subscribe
to receive more data or to unsubscribe from receiving data for certain
symbols.
Both requests are responded to by the Alpaca API with the current list
of subscriptions that the user has and we maintain this list in our
Subscription objects.