Introduce Source trait
The Alpaca real time data streaming API knows about two data sources:
IEX, providing data only from the investor's exchange, is available with
the free data plan and SIP, providing more accurate data, may only be
used in conjunction with the unlimited plan.
With this change we model these two possible sources, so that users can
specify them when establishing a connection. The way we do that is with
a trait that is implemented for two uninstantiable types. These types
will later be usable as generic type arguments to choose the source.
The main reason for modeling the problem space this way is that we want
to merely provide an ApiInfo object to the Subscribable::connect method
in the future and so we another channel for passing in the source to
use.