Introduce Unfold struct
The Alpaca real time data API does not send individual messages but
arrays of them. E.g., instead of sending one bar it sends a vector of
bars, containing one or more elements. That is arguably horrible to work
with and the reasoning for this design decision is questionable at best.
We do not want to expose this property to the user.
To "unfold" such inner arrays, this change introduces a stream wrapper,
Unfold, that emits those messages one by one, allowing us to hide this
implementation detail from users.