Introduce RealtimeData type
This change introduces the RealtimeData type, which implements
Subscribable. This type is the entity that users provide to the
Client::subscribe method in order to subscribe to real time market data.
Upon establishing a connection, the type takes care of authenticating
clients with the Alpaca API. The actual data subscription, however, has
to be performed by the user and can happen (and be changed) at any time
after the connection has been established.