Rename InsufficientFunds error variant to NotPermitted
In what can truly and uniquely be identified as Alpaca-style procedure,
the error code being reported when submitting orders with certain
parameters was silently changed from one value to another. Not only
that, the documentation was not adjusted at all, still mentioning the
old behavior.
To cope with this change, this patch does two things:
- it renames the InsufficientFunds error variant of the PostError and
  PatchError types in the api::v2::order module to NotPermitted
- it adjusts some of the tests that failed at certain times of the day
  to expect this new variant instead of InvalidInput, which was reported
  in the past