Make mock_stream function more generally usable
The mock_stream function that we use for testing the trade update
streaming code should in theory be usable for the testing of real time
market data update streaming as well. To make such usage possible, this
change parametrizes the function with the Subscribable instance to use
and moves it into the websocket module.