Remove data::v2::stocks module alias
The next release of the crate is bound to be backwards compatibility
breaking and requiring a minor version bump. As such, this change uses
this fact to get rid of the data::v2::stocks module alias that we only
kept around for the purpose of preserving backwards compatibility in the
past. Users should migrate to using data::v2::bars instead.