Introduce Subscription type
This change marks the beginning of a rework of the websocket
subscription functionality, which is currently used for streaming trade
updates. The rework will make the underlying mechanism more generally
usable and allow us to more easily implement streaming of current
quotes, trades, and bars in real time as well.
This change, specifically, introduces a new type, Subscription, that
later can be used to control certain aspects of the stream. It relies on
the websocket-util crate, which in turn provides a generic subscription
type that allows for sending and receiving "control messages" over an
associated websocket connection.