Rework GitLab CI script
When we last visited the stages and individual steps performed in CI,
GitLab did not yet support usage of multiple caches. That deficiency
greatly impeded what we were able to do and ultimately led to a
sub-optimal pipeline in many ways.
This change redesigns the CI script, greatly speeding up the pipeline in
the process. Specifically, we make the following changes:
- we separate building of debug and release targets
- we introduce four (smaller) caches for the various artifacts created
  (tarpaulin binary + instrumented obj files, obj files and incremental
  build artifacts for debug builds and release ones, and caching of the
  crates.io registry as stored in the Cargo home directory)
- we opt in to using fastzip for compression and decompression of
  artifacts and caches
- we no longer use verbose output when building and testing
- we remove the clean up of stale files, which did not actually work as
  it was supposed to because file access times were always set as part
  of cache extraction and so were never more than a few minutes in the
  past
- we collapse all tarpaulin related jobs, eliminating unnecessary and
  costly container image spin ups
- we solely rely on the tarpaulin job for testing; we must not have
  multiple test runs running in parallel or assumed invariants may be
  violated (e.g., when two tests check that they can create a watchlist
  of a certain name, but the Alpaca API allows for only one to exist)
- we remove the use of stages entirely; in fact, we even got rid of any
  all dependencies between jobs, to the point that all of them can start
  immediately and run concurrently

These changes allow us to get a no-op CI run (without any changes over a
previous one) done in just over three minutes. That is almost a ten fold
speed up over the ~30 minute runs that we have seen in the recent past.

Note that despite us removing the logic for cleaning up stale files,
doing so actually is still an unresolved issue at this time: we have not
yet found a good way to keep cache artifact size at bay over time. We
will have to evaluate how much of a problem that is moving forward and
find potential counter measures.