Bring over handshake tests to api::v2::updates module
We are in the process of reworking the mechanism backing the streaming
of trade updates over a websocket connection. Some of the type
definitions needed for the encoding and decoding of requests and
responses, respectively, are roughly similar to what we already have in
the handshake module and so it makes sense to adjust the existing tests
to work with them.
With this change we pull over said tests from the handshake module into
api::v2::updates, where the new functionality will reside, making
adjustments for accommodate differences along the way.