Fix link to coverage report
With the recent rework of the CI configuration we renamed, among other
things, the job taking care of capturing code coverage. Unfortunately,
the name of this job is embedded in the link to the code coverage HTML
that we provide in the README.
This change fixes said link by adjusting the name of the code coverage
job within.