Introduce Subscribable trait
At some point in the future the crate will support streaming of not only
trade updates but also other kinds of events over a websocket
connection. In the past we used the EventStream trait to abstract over
the two closely related streams we supported: trade updates and account
updates. Unfortunately, this trait was very much tailored to these two
use cases and is not general enough to support more or less arbitrary
event streaming.
To enable such use cases and also decouple the generic "subscribe" logic
in the Client type from potentially Alpaca API version specific details,
this change introduces a new trait that will form the abstraction
boundary: Subscribable. In a nutshell, this trait encapsulates the fact
that given some kind of initial input (e.g., credentials used for
authentication), a subscription and a stream are produced when
connecting. The subscription can be used to modify certain aspects of
the stream (e.g., to change the set of symbols for which to stream
historical aggregates) and the stream yields the actual consumable
elements.