Implement the Subscribable trait for TradeUpdates
This change implements the Subscribable trait that we introduced earlier
for our existing TradeUpdates type. In so doing we will later be able to
provide this type to a Client object in order to initiate streaming of
trade updates.
Note that unfortunately we need to make a couple of types public now,
because they appear in our public interface more or less by accident. We
could use Rust's "impl trait" for return values, but we decided against
that because it means that futures/streams effectively appear as unsized
in client code and cannot be named, necessitating boxing when that is
otherwise not a requirement.