Move remaining websocket streaming tests over
With the switch to using the new and shiny Subscribable trait for
implementing Client::subscribe, we are now ready to move the remaining
websocket stream related tests over, which exercise, among other things,
more of the new "handshake" code and associated error paths. Some of the
tests had to be adjusted slightly to accommodate for differing details
in the new implementation.