Correctly decode empty bar responses
As it turns out the Alpaca Data API does not strictly adhere to always
returning an array of bars in the response to a bars request. Rather, if
no data is returned it will just set the member to null. This value
currently causes a decode error, as serde_json does not treat that as a
valid variant for an array.
With this change we fix the problem by explicitly using the vec_from_str
function for decoding purposes. This function correctly handles such
upstream values.