Enable unreachable_pub lint
This change enables the unreachable_pub lint on the crate, warning us if
items declared as public are not actually accessible to client code
because they cannot be reached. To make this lint more meaningful in the
context of this crate, we "downgrade" a couple of items declared as pub
to pub(crate). Given that they were not accessible beforehand, these
changes do not change anything for client code.